feat: add tests, comments

+++ test/mocks/AddressArrayMock.sol
@@ -4,25 +4,24 @@ pragma solidity ^0.8.0;
 
 import "../../contracts/libraries/AddressArray.sol";
 
-
 contract AddressArrayMock {
     using AddressArray for AddressArray.Data;
 
     AddressArray.Data private _self;
 
-    function length() external view returns(uint256) {
+    function length() external view returns (uint256) {
         return AddressArray.length(_self);
     }
 
-    function at(uint i) external view returns(address) {
+    function at(uint256 i) external view returns (address) {
         return AddressArray.at(_self, i);
     }
 
-    function get() external view returns(address[] memory arr) {
+    function get() external view returns (address[] memory arr) {
         return AddressArray.get(_self);
     }
 
-    function push(address account) external returns(uint256) {
+    function push(address account) external returns (uint256) {
         return AddressArray.push(_self, account);
     }
